+++ src/common/imageio.c
@@ -494,14 +494,14 @@ void dt_imageio_to_fractional(float in, uint32_t *num, uint32_t *den)
 int dt_imageio_export(const uint32_t imgid, const char *filename, dt_imageio_module_format_t *format,
                       dt_imageio_module_data_t *format_params, const gboolean high_quality,
                       const gboolean copy_metadata, dt_imageio_module_storage_t *storage,
-                      dt_imageio_module_data_t *storage_params)
+                      dt_imageio_module_data_t *storage_params, int num, int total)
 {
   if(strcmp(format->mime(format_params), "x-copy") == 0)
     /* This is a just a copy, skip process and just export */
-    return format->write_image(format_params, filename, NULL, NULL, 0, imgid);
+    return format->write_image(format_params, filename, NULL, NULL, 0, imgid, num, total);
   else
     return dt_imageio_export_with_flags(imgid, filename, format, format_params, 0, 0, high_quality, 0, NULL,
-                                        copy_metadata, storage, storage_params);
+                                        copy_metadata, storage, storage_params, num, total);
 }
 
 // internal function: to avoid exif blob reading + 8-bit byteorder flag + high-quality override
@@ -511,7 +511,7 @@ int dt_imageio_export_with_flags(const uint32_t imgid, const char *filename,
                                  const gboolean high_quality, const int32_t thumbnail_export,
                                  const char *filter, const gboolean copy_metadata,
                                  dt_imageio_module_storage_t *storage,
-                                 dt_imageio_module_data_t *storage_params)
+                                 dt_imageio_module_data_t *storage_params, int num, int total)
 {
   dt_develop_t dev;
   dt_dev_init(&dev, 0);
@@ -803,11 +803,11 @@ int dt_imageio_export_with_flags(const uint32_t imgid, const char *filename,
     // last param is dng mode, it's false here
     length = dt_exif_read_blob(exif_profile, pathname, imgid, sRGB, processed_width, processed_height, 0);
 
-    res = format->write_image(format_params, filename, outbuf, exif_profile, length, imgid);
+    res = format->write_image(format_params, filename, outbuf, exif_profile, length, imgid, num, total);
   }
   else
   {
-    res = format->write_image(format_params, filename, outbuf, NULL, 0, imgid);
+    res = format->write_image(format_params, filename, outbuf, NULL, 0, imgid, num, total);
   }
 
   dt_dev_pixelpipe_cleanup(&pipe);
