+++ src/common/imageio.h
@@ -57,7 +57,7 @@ struct dt_imageio_module_data_t;
 int dt_imageio_export(const uint32_t imgid, const char *filename, struct dt_imageio_module_format_t *format,
                       struct dt_imageio_module_data_t *format_params, const gboolean high_quality,
                       const gboolean copy_metadata, dt_imageio_module_storage_t *storage,
-                      dt_imageio_module_data_t *storage_params);
+                      dt_imageio_module_data_t *storage_params, int num, int total);
 
 int dt_imageio_export_with_flags(const uint32_t imgid, const char *filename,
                                  struct dt_imageio_module_format_t *format,
@@ -65,7 +65,7 @@ int dt_imageio_export_with_flags(const uint32_t imgid, const char *filename,
                                  const int32_t display_byteorder, const gboolean high_quality,
                                  const int32_t thumbnail_export, const char *filter,
                                  const gboolean copy_metadata, dt_imageio_module_storage_t *storage,
-                                 dt_imageio_module_data_t *storage_params);
+                                 dt_imageio_module_data_t *storage_params, int num, int total);
 
 size_t dt_imageio_write_pos(int i, int j, int wd, int ht, float fwd, float fht,
                             dt_image_orientation_t orientation);
