+++ src/imageio/format/ppm.c
@@ -34,7 +34,7 @@ void cleanup(dt_imageio_module_format_t *self)
 }
 
 int write_image(dt_imageio_module_data_t *ppm, const char *filename, const void *in_tmp, void *exif,
-                int exif_len, int imgid)
+                int exif_len, int imgid, int num, int total)
 {
   const uint16_t *in = (const uint16_t *)in_tmp;
   int status = 0;
