+++ src/imageio/storage/CMakeLists.txt
@@ -1,7 +1,7 @@
 cmake_minimum_required(VERSION 2.6)
 
 include_directories("${CMAKE_CURRENT_BINARY_DIR}/../../" "${CMAKE_CURRENT_SOURCE_DIR}")
-set(MODULES disk email gallery latex pdf)
+set(MODULES disk email gallery latex)
 
 if(FLICKCURL_FOUND)
 	list(APPEND MODULES flickr)
