+++ src/imageio/storage/disk.c
@@ -283,7 +283,7 @@ int store(dt_imageio_module_storage_t *self, dt_imageio_module_data_t *sdata, co
   if(fail) return 1;
 
   /* export image to file */
-  if(dt_imageio_export(imgid, filename, format, fdata, high_quality, TRUE, self, sdata) != 0)
+  if(dt_imageio_export(imgid, filename, format, fdata, high_quality, TRUE, self, sdata, num, total) != 0)
   {
     fprintf(stderr, "[imageio_storage_disk] could not export to file: `%s'!\n", filename);
     dt_control_log(_("could not export to file `%s'!"), filename);
