+++ src/imageio/storage/email.c
@@ -127,7 +127,7 @@ int store(dt_imageio_module_storage_t *self, dt_imageio_module_data_t *sdata, co
 
   attachment->file = g_build_filename(tmpdir, dirname, (char *)NULL);
 
-  if(dt_imageio_export(imgid, attachment->file, format, fdata, high_quality, FALSE, self, sdata) != 0)
+  if(dt_imageio_export(imgid, attachment->file, format, fdata, high_quality, FALSE, self, sdata, num, total) != 0)
   {
     fprintf(stderr, "[imageio_storage_email] could not export to file: `%s'!\n", attachment->file);
     dt_control_log(_("could not export to file `%s'!"), attachment->file);
