+++ src/imageio/storage/latex.c
@@ -338,7 +338,7 @@ int store(dt_imageio_module_storage_t *self, dt_imageio_module_data_t *sdata, co
   dt_pthread_mutex_unlock(&darktable.plugin_threadsafe);
 
   /* export image to file */
-  dt_imageio_export(imgid, filename, format, fdata, high_quality, FALSE, self, sdata);
+  dt_imageio_export(imgid, filename, format, fdata, high_quality, FALSE, self, sdata, num, total);
 
   printf("[export_job] exported to `%s'\n", filename);
   char *trunc = filename + strlen(filename) - 32;
